requirement.txt
CTkMessagebox==2.5
customtkinter==5.2.2
darkdetect==0.8.0
future==1.0.0
greenlet==3.0.3
packaging==24.1
pillow==10.3.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
setuptools==70.0.0
six==1.16.0
SQLAlchemy==2.0.30
tkintertable==1.3.3
typing_extensions==4.12.2
